+++ Muxy/Views/Sidebar/ExpandedProjectRow.swift
@@ -255,8 +255,8 @@ struct ExpandedProjectRow: View {
 
     @ViewBuilder
     private var iconOrBadge: some View {
-        if showShortcutBadge, let shortcutIndex {
-            ShortcutIconBadge(number: shortcutIndex, size: UIMetrics.iconXXL)
+        if let shortcutIndex, let hint = shortcutHint {
+            ShortcutIconBadge(number: shortcutIndex, size: UIMetrics.iconXXL, combo: hint)
         } else {
             projectIcon
         }
@@ -367,13 +367,11 @@ struct ExpandedProjectRow: View {
         return AnyShapeStyle(Color.clear)
     }
 
-    private var showShortcutBadge: Bool {
+    private var shortcutHint: KeyCombo? {
         guard let shortcutIndex,
               let action = ShortcutAction.projectAction(for: shortcutIndex)
-        else { return false }
-        return ModifierKeyMonitor.shared.isHolding(
-            modifiers: KeyBindingStore.shared.combo(for: action).modifiers
-        )
+        else { return nil }
+        return ModifierKeyMonitor.shared.hint(for: action)
     }
 
     private func pickLogoImage() {
